Add files via upload
This is the beginnings of a program to help people learn to master classic memory training techniques.  This is just the first of several modules and allows the user to develop speed at translating the numbers 0-100 into word-pegs.  When a person gets good at that they can retain phone numbers, flight numbers - any kind of numbers more quickly and more permanently.